Global envelope test for spatio-temporal separability using S-function

Performs a global envelope test of the null hypothesis of first-order separability for a spatio-temporal point process. The observed separability diagnostics \(S(u,t)\), \(S_{\mathrm{space}}(u)\) and/or \(S_{\mathrm{time}}(t)\) are compared to a reference distribution obtained from permuted versions of the data.

global.envelope.test(
  X,
  sim.procedure = c("pure_per", "block_per"),
  nsim = 25L,
  nblocks = 5L,
  nperm = 199L,
  n.grid = c(20L, 20L, 10L),
  s.region = matrix(c(0, 0, 1, 0, 1, 1, 0, 1), ncol = 2, byrow = TRUE),
  t.region = c(0, 1),
  owin = NULL,
  eps = NULL,
  del = NULL,
  tests = c("S.test", "S.space.test", "S.time.test"),
  GET.args = NULL
)

Arguments

X

A numeric matrix or data frame with at least three columns giving \((x,y,t)\).

sim.procedure

Character string specifying the permutation strategy: "pure_per" or "block_per".

nsim

Integer. Number of permutations for "pure_per".

nblocks

Integer (>= 2). Number of temporal blocks for block permutation. Used only for "block_per".

nperm

Integer. Number of block permutations for "block_per".

n.grid

Integer/numeric vector of length 3 specifying grid resolution in \(x,y,t\).

s.region

Numeric matrix with two columns specifying polygon vertices of the spatial window.

t.region

Numeric vector of length 2 specifying temporal window c(tmin,tmax).

owin

Optional window of class "owin" (from spatstat.geom). If supplied, values outside the window are set to NA (pixels mode).

eps

Optional numeric scalar (>0). Spatial bandwidth. If NULL, estimated internally.

del

Optional numeric scalar (>0). Temporal bandwidth. If NULL, estimated internally.

tests

Character vector indicating which diagnostics to test. Any of "S.test", "S.space.test", "S.time.test".

GET.args

Optional named list of extra arguments passed to global_envelope_test (e.g. alternative, savefuns, nstep).

Value

A list with components:

Bandwidth_s

Spatial bandwidth used.

Bandwidth_t

Temporal bandwidth used.

S.test, S.space.test, S.time.test

For each requested test: p-values for ERL and AREA envelopes, plus optional plots if created.

Details

Two permutation strategies are supported:

"pure_per"

Pure permutation: randomly permutes the time coordinates.

"block_per"

Block permutation: permutes time in blocks to preserve short-range temporal dependence.

The GET package is used to construct global envelopes and compute p-values.

The null hypothesis is $$H_0:\ \rho(u,t)=\rho_{\mathrm{space}}(u)\rho_{\mathrm{time}}(t).$$ The function computes the chosen diagnostics using S.based.functions on a pixel grid (at="pixels") and applies global_envelope_test.

To keep curve lengths identical (required by GET), any NA values induced by an owin mask are removed using the same indices for the observed and all simulated curves.

References

Ghorbani, M., Vafaei, N., Dvořák, J., and Myllymäki, M. (2021). Testing the first-order separability hypothesis for spatio-temporal point patterns. Computational Statistics & Data Analysis, 161, 107245.

See also

Author

Nafiseh Vafaei nafiseh.vafaei@slu.se
Mohammad Ghorbani mohammad.ghorbani@slu.se

Examples

# \donttest{
if (requireNamespace("GET", quietly = TRUE)) {
  set.seed(123)
  X <- cbind(stats::runif(100), stats::runif(100), stats::runif(100, 0, 1))
  s.region <- matrix(c(0,0, 1,0, 1,1, 0,1), ncol = 2, byrow = TRUE)
  t.region <- c(0, 1)

  res <- global.envelope.test(
    X = X,
    sim.procedure = "pure_per",
    nsim = 19,
    n.grid = c(10,10,10),
    s.region = s.region,
    t.region = t.region,
    tests = c("S.test","S.time.test")
  )
  str(res)
}
#> List of 4
#>  $ Bandwidth_s: num 0.0612
#>  $ Bandwidth_t: num 0.107
#>  $ S.test     :List of 4
#>   ..$ p.value.erl  : num 0.4
#>   ..$ p.value.area : num 0.45
#>   ..$ envelope.erl :Classes 'global_envelope' and 'data.frame':	1000 obs. of  5 variables:
#>   .. ..$ r      : int [1:1000] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ obs    : num [1:1000] 0.1921 0.0668 0.2546 1.4428 1.3652 ...
#>   .. ..$ central: num [1:1000] 0.938 0.959 0.964 1.246 1.112 ...
#>   .. ..$ lo     : num [1:1000] 0.000472 0.007196 0.035356 0.137683 0.024095 ...
#>   .. ..$ hi     : num [1:1000] 4.92 5.19 5.81 4.29 3.91 ...
#>   .. ..- attr(*, "xlab")=  expression(italic(r))
#>   .. ..- attr(*, "ylab")=  expression(italic(T(r)))
#>   .. ..- attr(*, "alternative")= chr "two.sided"
#>   .. ..- attr(*, "method")= chr "Global envelope test"
#>   .. ..- attr(*, "type")= chr "erl"
#>   .. ..- attr(*, "alpha")= num 0.05
#>   .. ..- attr(*, "M")= Named num [1:20] 0.4 0.35 0.8 0.5 0.3 0.05 0.1 0.55 0.2 0.65 ...
#>   .. .. ..- attr(*, "names")= chr [1:20] "obs" "sim1" "sim2" "sim3" ...
#>   .. ..- attr(*, "M_alpha")= Named num 0.1
#>   .. .. ..- attr(*, "names")= chr "sim6"
#>   .. ..- attr(*, "call")= language individual_global_envelope_test(curve_set = curve_sets, type = type, alpha = alpha,      alternative = alternativ| __truncated__ ...
#>   .. ..- attr(*, "p")= num 0.4
#>   ..$ envelope.area:Classes 'global_envelope' and 'data.frame':	1000 obs. of  5 variables:
#>   .. ..$ r      : int [1:1000] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ obs    : num [1:1000] 0.1921 0.0668 0.2546 1.4428 1.3652 ...
#>   .. ..$ central: num [1:1000] 0.938 0.959 0.964 1.246 1.112 ...
#>   .. ..$ lo     : num [1:1000] 0.000472 0.007196 0.035356 0.137683 0.024095 ...
#>   .. ..$ hi     : num [1:1000] 4.92 5.19 5.81 4.29 3.91 ...
#>   .. ..- attr(*, "xlab")=  expression(italic(r))
#>   .. ..- attr(*, "ylab")=  expression(italic(T(r)))
#>   .. ..- attr(*, "alternative")= chr "two.sided"
#>   .. ..- attr(*, "method")= chr "Global envelope test"
#>   .. ..- attr(*, "type")= chr "area"
#>   .. ..- attr(*, "alpha")= num 0.05
#>   .. ..- attr(*, "M")= Named num [1:20] 0.0495 0.0494 0.0496 0.0495 0.0495 ...
#>   .. .. ..- attr(*, "names")= chr [1:20] "obs" "sim1" "sim2" "sim3" ...
#>   .. ..- attr(*, "M_alpha")= Named num 0.0492
#>   .. .. ..- attr(*, "names")= chr "sim6"
#>   .. ..- attr(*, "call")= language individual_global_envelope_test(curve_set = curve_sets, type = type, alpha = alpha,      alternative = alternativ| __truncated__ ...
#>   .. ..- attr(*, "p")= num 0.45
#>  $ S.time.test:List of 4
#>   ..$ p.value.erl  : num 0.15
#>   ..$ p.value.area : num 0.25
#>   ..$ envelope.erl :Classes 'global_envelope' and 'data.frame':	10 obs. of  5 variables:
#>   .. ..$ r      : int [1:10] 1 2 3 4 5 6 7 8 9 10
#>   .. ..$ obs    : num [1:10] 0.913 0.946 1.003 1.034 1.005 ...
#>   .. ..$ central: num [1:10] 1.002 0.993 0.995 0.995 0.988 ...
#>   .. ..$ lo     : num [1:10] 0.913 0.922 0.921 0.91 0.923 ...
#>   .. ..$ hi     : num [1:10] 1.08 1.05 1.07 1.06 1.06 ...
#>   .. ..- attr(*, "xlab")=  expression(italic(r))
#>   .. ..- attr(*, "ylab")=  expression(italic(T(r)))
#>   .. ..- attr(*, "alternative")= chr "two.sided"
#>   .. ..- attr(*, "method")= chr "Global envelope test"
#>   .. ..- attr(*, "type")= chr "erl"
#>   .. ..- attr(*, "alpha")= num 0.05
#>   .. ..- attr(*, "M")= Named num [1:20] 0.15 0.1 0.05 0.75 1 0.7 0.6 0.9 0.2 0.55 ...
#>   .. .. ..- attr(*, "names")= chr [1:20] "obs" "sim1" "sim2" "sim3" ...
#>   .. ..- attr(*, "M_alpha")= Named num 0.1
#>   .. .. ..- attr(*, "names")= chr "sim1"
#>   .. ..- attr(*, "call")= language individual_global_envelope_test(curve_set = curve_sets, type = type, alpha = alpha,      alternative = alternativ| __truncated__ ...
#>   .. ..- attr(*, "p")= num 0.15
#>   ..$ envelope.area:Classes 'global_envelope' and 'data.frame':	10 obs. of  5 variables:
#>   .. ..$ r      : int [1:10] 1 2 3 4 5 6 7 8 9 10
#>   .. ..$ obs    : num [1:10] 0.913 0.946 1.003 1.034 1.005 ...
#>   .. ..$ central: num [1:10] 1.002 0.993 0.995 0.995 0.988 ...
#>   .. ..$ lo     : num [1:10] 0.913 0.922 0.921 0.91 0.923 ...
#>   .. ..$ hi     : num [1:10] 1.08 1.05 1.07 1.06 1.06 ...
#>   .. ..- attr(*, "xlab")=  expression(italic(r))
#>   .. ..- attr(*, "ylab")=  expression(italic(T(r)))
#>   .. ..- attr(*, "alternative")= chr "two.sided"
#>   .. ..- attr(*, "method")= chr "Global envelope test"
#>   .. ..- attr(*, "type")= chr "area"
#>   .. ..- attr(*, "alpha")= num 0.05
#>   .. ..- attr(*, "M")= Named num [1:20] 0.0485 0.0489 0.0427 0.1449 0.1919 ...
#>   .. .. ..- attr(*, "names")= chr [1:20] "obs" "sim1" "sim2" "sim3" ...
#>   .. ..- attr(*, "M_alpha")= Named num 0.0477
#>   .. .. ..- attr(*, "names")= chr "sim12"
#>   .. ..- attr(*, "call")= language individual_global_envelope_test(curve_set = curve_sets, type = type, alpha = alpha,      alternative = alternativ| __truncated__ ...
#>   .. ..- attr(*, "p")= num 0.25
# }